Update gopsutil dependency to enable 32-bit builds

+++ Godeps/_workspace/src/github.com/shirou/gopsutil/disk/disk_linux.go
@@ -318,7 +318,7 @@ func GetDiskSerialNumber(name string) string {
 }
 
 func getFsType(stat syscall.Statfs_t) string {
-	t := stat.Type
+	t := int64(stat.Type)
 	ret, ok := fsTypeMap[t]
 	if !ok {
 		return ""
